First page (summary report) is now up and running.
This is the Angular version of previous gnatquilt
(which was developped with Google Closure Library).
Let's call it gnatquilt-angular.

Indexing page should be working, with some few minors
(project information can't be collapsed as of right now).
I may add some filters on the coverage status, it might be
something useful.

The app.component is the entry point of the application, which
is splitted in different components, roughly being:
  - The report.component which gives the layout of the whole page
   (Overview, information for each project, general coverage information).
  - The source_table related components.
    They are used to display per-project metrics / per-report metrics.
    (coverage for each project, coverage for the overview).

Each component usually has a typescript file (for data processing,
event reaction), an HTML file (structure the report and set the different
components to be instantiated) and possibly a style sheet.

I will add comments later, type annotations are usually quite informative.
Best IDE to run and test is probably Webstorm, you can get a free license
for 30 days if you want to take a shot at executing the application.

For now, mock data is being used (a hard-encoded JSON file in the sources).

TN: TA14-003

Change-Id: I1ba2f830f095fd96b0fec9e0d32a940eaea766fd