Routes integration and services refactoring
To load another HTML page, or change the loaded components to
add another view, we need to use routes. Routes associate to an
url a component and instantiate this component where the <router-outlet>
tag is found (most of the time, in the app.component template,
which is the parent component of the application).
This mapping can be found in src/app/route/routing.module.ts
Given this mapping (to which url link which component),
we need to set routerLinks in the application
(references to other url in the indexed report that will enable navigation).
In GNATcoverage indexed report, those references
will be links to the annotated source file report.
To avoid performance issues due to reinstatiation of components
when navigating back to an already visited url, we also need to set
a custom route strategy, to reuse components when needed.

Only problem with routes is when we need to pass info to
the route destination component.
The only way to do so is to pass string parameters,
that will either be encoded in the url or given as extra to the route.
Complex objects will thus be stringified,
and this might cause performance issues.
For that reason, and also for clarification purposes,
I refactored the services (I was not using services before, purely classes)
and used dependency injection to provide complex information
that is needed in the whole application (such as the
status that are of interest, if there are exemptions or not).

Change-Id: Iaf2800a150f699f044a2cb9f1d32e781280b7982
TN: TA14-003