integrated annotated source component
The other important thing about the report
is the annotated source view, which gives
thorough coverage information (top level
information, whether a line has been covered
or not, and detailed information when a line
is partially covered / not coverable).
This view still needs to be enriched, especially
when there is machine code attached as coverage
information.
For now we have the general structure of gnatquilt,
and further testing will be done when integrating
with gnatcov, to enhance the angular version.

The annotated source view, as mock data is still
of use, only functions for `pk1.adb` (there is data
only for this file).

I also corrected the type annotations all along, and removed
inferrable types to have fully type annotated variables.

TN: TA14-003

Change-Id: I0732ac25a7afda8f96d2b03fd3e7585dfcdd4ef8