Integration with GNATcoverage
As we are aiming at being maximally consistent with GNATquilt previous
use, integration with GNATcoverage has to be as similar as possible.

For reminders, GNATcoverage was issuing several files for the dhtml
report (which is where GNATquilt is used).
These files consisted of:
- An index.html file, which can be opened in the browser  and enables
  the user to navigate throughout the general report and the files.
- JS files for every annotated source

The main behaviour here is to have a static html file that can be open
without the need of any server or data being served dynamically.
We also want to preserve that "split by file" behaviour, that allows
better performance on larger projects, as the loaded chunks for each
annotated file can be quite big.

To do so, we know that one file will always be generated: the report file,
that contains aggregated stats and information on the sources.
We will assume in this commit that this file is named 'report.js'.
This file will just populate a JS variable, that we will assume to be
externally defined in GNATquilt.
It will contain properties, and the other JS generated by GNATquilt (for
each annotated source) will also populate this variable with adequate p
roperty names and adequate information.

TN: TA14-003
Change-Id: I49063b962506d43784483aaa0d5971e5e83dbb06