Enhance GNATquilt performance
GNATcoverage being an industrial tool, projects loaded in GNATquilt can
be quite big and project file can very well be tens of thousands of lines.

In GNATquilt previous version, as the generated code was very optimized
(because of the Closure Compiler that generates very efficient code), this
issue was mitigated but bigger files were still taking some seconds to
load. In the Angular version, the loading time was approximately doubled,
so optimizations were needed.

We make use of a node package for virtual scrolling, not in the standard
Angular library for two reasons:
- Variable sized items are not supported (it is an
  experimental feature).
- Parent scroll can't be used

The performance is OK. There are still some tedious issues with the
scrolling that need to be fixed though.
Among them being:
- Invalidating cache measurement when expanding a line sometimes causes
  a "jump" issue.
- Resizing the frame makes the annotated lines expand out of their div
  container.
- This last issue could be solved by surrounding, as it would normally be
  done, the table rows of the annotated source (i.e. the source line) with
  HTML table and tbody tags, but for unknown reasons, the virtual scroller
  then bugs out the display.
  For this, see the source-file component in source/source-file, where the
  virtual scroller is used.

Also reformatted files.
TN: TA14-003
Change-Id: I482f50cfaba2aba34f5b4015305b9a36b2916561