Fixed-width font for messages
GNATcoverage has a special option --show-mcdc-vector to get more
information about the MD/DC evaluation vectors. It now uses "-" for non
evaluated condition in a given evaluation. GNATquilt must be modified in
order to have the right alignment for MC/DC vectors. To do that, we will
use a fixed-width font for messages.

Change-Id: Ide12ce6d56a08b2f04c745a6d5fb2f913229b1e9
TN: TB27-018