Fix display bug in stmt+decision mode
A decision SCO does not necessarily have condition SCO(s) associated.
GNATquilt implicitely made that assumption, which resulted in a bogus
execution in the browser. The observed behaviour is that when in
stmt+decision mode and in a file containing decisions, the SCO text is not
printed in the violation.

Change-Id: I6c2c4def7a854cc88d382394abb11d39eb823bed