Support entity-based reporting
With integration of entity-based capabilities reporting in gnatcov, we
enhance the dhtml output to provide this information in a convenient
and simple way to the user.

This also revamps some bits of GNATquilt design, notably how the context
(storing general-purpose information, such as the coverage levels
enabled etc) is computed and passed on to the various views (to avoid
recomputing it every time it is needed).

TN: S213-006
Depends-On: I5c4b8aad44a572d759878199d4248970c2ff0bfa
Change-Id: Ib64bc421d7615c7820415d9f9c4cd5c9d2e7e2ba