Add new coverage state for non-instrumented entities
This change adds support for non instrumented entites. If the
entity stats concerning non instrumented entites are present in
the .hunk.js files generated by gnatcov these will always be
displayed in the report. It is thus the responsibility of gnatcov
to decide wether these stats should be present or not in the
report.

TN: V419-003
Change-Id: I89d9bb4f734eb68fcf9b7221b3322820ab1f6add