Introduce support for scoped metrics
This commit introduces support for scoped metrics. Such metrics are
rendered in the source file component, in the header part.

They are presented in a nested fashion: the user can expand the list of
top-level scoped metrics, and from there, expand (for each scope) its
nested scoped metrics.

The user can also click on a scope name to jump to its definition point
in the report.

Change-Id: I5f6c67adc51b68466cae85168163767fd7a2a7d5
TN: R417-010