Modify test project to be able to test scrolling behavior
This commit adds a function and artificially increase the size (in line
numbers) of a package, to be able to make use of scrolling features of
the application. This is useful to implement a test that checks against
subprogram metrics and navigation to a subprogram from these metrics. This
test is implemented in a posterior commit.

The commit also makes the necessary baseline adaptations as coverage
statistics change.

Also replace the code duplication with a symbolic link to the testsuite
project.

Change-Id: Idd001992df544bb6e30e76d638d101d71b27566b
TN: R417-010