Load the source file report only after the report has been loaded
On some occasions, the report script was reloaded after the source file
script, which resulted in an `undefined REPORT` error, as the `REPORT`
variable is defined in the `report.js` script. As a reminder, the
`REPORT` variable is a javascript variable that is used to fill report /
source file coverage information.

To avoid this kind of situation, we make sure to wait for the report
script to load prior to loading the source file script.