Remove symbolic link
Remove the test symbolic link that was pointing to a test inside the
testsuite. As a reminder, the `test/dhtml` is supposed to contain
gnatcov artifacts that are used to run a development server. Most of the
time, it is more convenient to setup dedicated artifacts for the problem
that we are investigating, and the symbolic link becomes a burden.