Fix regression when using entity reporting
To make sure to load the source after the report has been loaded, we
subscribe to the report and emit the source when the subscription gets
fulfilled. We want to do that only when the report loads, not when it
potentially changes afterwards, so make this a 1-time subscription, to
avoid emitting outdated sources versions.