Override the search functionality of the browser
As gnatquilt uses virtual scrolling, it renders a portion of the source
code at a time, for performance purposes. This has the downside of
breaking the native search functionality. Reimplement it to have the
expected search behavior (i.e. look for strings in the whole file, rather
than only the visible portion of it).